refactor: restructure providers directory to group backends
