--- merino/backends/__init__.py
@@ -1,1 +1,0 @@
-"""Merino backends."""
