--- merino/backends/exceptions.py
@@ -1,5 +1,0 @@
-"""Backend module exceptions."""
-
-
-class BackendError(Exception):
-    """Error specific to backend functions."""
