+++ merino/exceptions.py
@@ -1,6 +1,10 @@
 """Merino specific exceptions."""
 
 
+class BackendError(Exception):
+    """Error specific to provider backend functions."""
+
+
 class InvalidProviderError(Exception):
     """Raised when an unknown provider encountered."""
