+++ merino/providers/adm/__init__.py
@@ -1,0 +1,1 @@
+"""AdM provider"""
