+++ merino/providers/adm/backends/__init__.py
@@ -1,0 +1,1 @@
+"""AdM backends"""
