+++ merino/providers/weather/__init__.py
@@ -1,0 +1,1 @@
+"""Weather provider"""
