+++ merino/providers/weather/backends/__init__.py
@@ -1,0 +1,1 @@
+"""Weather backends"""
