+++ tests/unit/providers/weather/test_weather.py
@@ -10,11 +10,11 @@ import pytest
 from pytest import LogCaptureFixture
 from pytest_mock import MockerFixture
 
-from merino.backends.exceptions import BackendError
 from merino.config import settings
+from merino.exceptions import BackendError
 from merino.middleware.geolocation import Location
 from merino.providers.base import BaseSuggestion, SuggestionRequest
-from merino.providers.weather import (
+from merino.providers.weather.weather import (
     CurrentConditions,
     Forecast,
     Provider,
@@ -159,7 +159,7 @@ async def test_query_error(
     assert suggestions == expected_suggestions
     actual_log_messages: list[dict[str, str]] = [
         {"levelname": record.levelname, "message": record.message}
-        for record in filter_caplog(caplog.records, "merino.providers.weather")
+        for record in filter_caplog(caplog.records, "merino.providers.weather.weather")
     ]
     assert actual_log_messages == expected_log_messages
